Player Hit Delay Is Now Consistent
This version is just in case I massively mess up what I'm about to
attempt, and need to revert back.
What I aim to fix in the next version is the occasional
NullReferenceException, the inconsistent player damage, and hopefully I
can manage to move all IE numerators to an object that will never be
destroyed, as opposed to where they are currently located: the enemy.